update logger request - handle order api controller - handle order request - handle order response - add order item qr code link attribute - add taxonomy name attribute to product - cast double to total amount of order - handle v1 and v2 for order - add order items resourse - handle firebase dynamic link service

+++ app/Services/OrderService.php
@@ -59,7 +59,7 @@ class OrderService {
 
             foreach ($keyedProducts as $vendorId => $vendorProducts) {
                 $data['vendor_id'] = $vendorId;
-                $data['total'] = array_sum(array_map(function ($product) use ($data) {
+                $data['total'] = (double) array_sum(array_map(function ($product) use ($data) {
                     $inputItem = Arr::first($data['products'], function ($value, $key) use ($product) {
                         return (int)($value['id']) === $product->id;
                     });
